reporting/streamlit/requirements.txt
streamlit
pandas
numpy
matplotlib
plotly
duckdb
google-cloud-secret-manager
requests
streamlit-lottie
network
vertexai
google-cloud-aiplatform
python-dotenv